improve validation for user
